+++ saleor/graphql/account/mutations/staff/customer_create.py
@@ -1,7 +1,8 @@
 from urllib.parse import urlencode
 
 from django.contrib.auth.tokens import default_token_generator
-from django.forms import model_to_dict
+from django.core.exceptions import ValidationError
+from django.db import IntegrityError, transaction
 
 from .....account import events as account_events
 from .....account import models
@@ -65,10 +66,43 @@ class CustomerCreate(BaseCustomerCreate):
 
         instance.search_document = prepare_user_search_document_value(instance)
 
-        (instance, created) = models.User.objects.get_or_create(
-            email=instance.email, defaults=model_to_dict(instance)
-        )
-        # TODO - it never creates because it was created earlier
+        """
+        Prevent race condition when saving customer.
+
+        This is a hacky solution that catches DB error, to deduct that entity was created before.
+
+        Normally we would use get_or_create here, but it will lose reference - get_or_create
+        returns new instance of model, but original "instance" is being returned in mutation.
+
+        This limitation comes from the Base/Model mutation. We can solve this e.g. by returning
+        actual instance from save()
+
+        https://linear.app/saleor/issue/EXT-2162
+        """
+        try:
+            with transaction.atomic():
+                instance.save()
+        except IntegrityError:
+            try:
+                # Verify if object already exists in DB.
+                # If yes, it means we have a race-condition
+                # This eventually leads to ValidationError because this user
+                # already exists
+                models.User.objects.get(email=instance.email)
+
+                raise ValidationError(
+                    {
+                        # This validation error mimics built-in validation error
+                        # So graphQL response is the same
+                        "email": ValidationError(
+                            "User with this Email already exists.",
+                            code=AccountErrorCode.UNIQUE.value,
+                        )
+                    }
+                )
+            except instance.DoesNotExist:
+                pass
+            raise
 
         cls.call_event(manager.customer_created, instance)
         account_events.customer_account_created_event(user=instance)